requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe @ file:///opt/concourse/worker/volumes/live/cb778296-98db-45ad-411e-6f726e102dc3/volume/markupsafe_1594371638608/work
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
twilio==6.45.3
urllib3==1.25.10
Werkzeug==1.0.1
